refactor: split lint cli args and app flow
